moved the statistics command back to a list command under the 'list' group
