+++ praetorian_cli/handlers/cli_decorators.py
@@ -66,5 +66,5 @@ def list_params(filter_by, has_details=True):
 def pagination(func):
     func = click.option('-o', '--offset', default='', help='List results from an offset')(func)
     func = click.option('-p', '--page', type=click.Choice(('first', 'all')), default='first',
-                        help='Pagination mode. "all" pages up to 1000 pages.', show_default=True)(func)
+                        help='Pagination mode. "all" pages up to 10,000 pages.', show_default=True)(func)
     return func
