+++ praetorian_cli/handlers/list.py
@@ -1,7 +1,7 @@
 import click
 
 from praetorian_cli.handlers.chariot import chariot
-from praetorian_cli.handlers.cli_decorators import list_params
+from praetorian_cli.handlers.cli_decorators import list_params, pagination, cli_handler
 from praetorian_cli.handlers.utils import render_offset, render_list_results, pagination_size, error
 
 
@@ -179,3 +179,53 @@ def seeds(chariot, type, filter, details, offset, page):
         error('When the DNS filter is specified, you also need to specify the type of the filter: ip or domain.')
 
     render_list_results(chariot.seeds.list(type, filter, offset, pagination_size(page)), details)
+
+
+@list.command
+@click.option('-f', '--filter', default='', help='Filter by statistic type or name')
+@click.option('--from', 'from_date', help='Start date (YYYY-MM-DD)')
+@click.option('--to', 'to_date', help='End date (YYYY-MM-DD)')
+@click.option('-d', '--details', is_flag=True, default=False, help='Show detailed information')
+@click.option('--help-stats', is_flag=True, help='Show detailed information about statistic types')
+@pagination
+@cli_handler
+def statistics(chariot, filter, from_date, to_date, details, offset, page, help_stats):
+    """ List statistics
+    Retrieve and display a list of statistics with optional date range filtering.
+    Use --help-stats for detailed information about available statistic types.
+
+    \b
+    Example usages:
+        - praetorian chariot list statistics
+        - praetorian chariot list statistics --filter "my#status"
+        - praetorian chariot list statistics --from 2024-12-01 --to 2024-12-31
+        - praetorian chariot list statistics --details
+        - praetorian chariot list statistics --page all
+        - praetorian chariot list statistics --help-stats
+    """
+    if help_stats:
+        click.echo(chariot.statistics.util.get_statistics_help())
+        return
+
+    # Map common filter aliases to StatsFilter values
+    filter_map = {
+        'risks': chariot.statistics.util.RISKS,
+        'risk_events': chariot.statistics.util.RISK_EVENTS,
+        'assets_by_status': chariot.statistics.util.ASSETS_BY_STATUS,
+        'assets_by_class': chariot.statistics.util.ASSETS_BY_CLASS,
+        'seeds': chariot.statistics.util.SEEDS
+    }
+
+    # Use mapped filter if available, otherwise use raw filter string
+    actual_filter = filter_map.get(filter, filter)
+
+    render_list_results(
+        chariot.statistics.list(
+            actual_filter,
+            from_date,
+            to_date,
+            offset,
+            pagination_size(page)
+        ),
+        details
+    )
